Move serde stuff to module
