+++ src/de/mod.rs
@@ -1,0 +1,2 @@
+pub mod variant_de;
+pub mod wbem_class_de;
